requirements.txt
# informal dependencies list for scheme.org related modules

networkx
rdflib
tabletext
graphviz
jsonschema
orderedset
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
